prepareTrack function of Kalman w/ RefTrack. Some other smaller stuff
git-svn-id: svn+ssh://svn.code.sf.net/p/genfit/code/branches/genfit2@1094 4ab16672-d4cc-4d65-938a-feac61c82957

+++ core/src/AbsTrackRep.cpp
@@ -72,7 +72,7 @@ void AbsTrackRep::get6DStateCov(const MeasuredStateOnPlane* stateInput, TVectorD
 
 
 void AbsTrackRep::Print(const Option_t*) const {
-  std::cout << "genfit::TrackRep, pdgCode = " << pdgCode_ << ". PropDir = " << propDir_ << "\n";
+  std::cout << "genfit::TrackRep, pdgCode = " << pdgCode_ << ". PropDir = " << (int)propDir_ << "\n";
 }
